// src/types.ts
export type Player = {
    id: number;
    name: string;
    iconClass: string;
    colorClass: string;
};

export type Move = {
    squareId: number;
    player: Player;
};

export type GameStatus = {
    isComplete: boolean;
    winner: Player | null;
};

export type Game = {
    moves: Move[];
    status: GameStatus;
};

export type GameState = {
    currentGameMoves: Move[];
    history: {
        currentRoundGames: Game[];
        allGames: Game[];
    };
};

export type DerivedGame = {
    moves: Move[];
    currentPlayer: Player;
    status: GameStatus;
};

export type DerivedStats = {
    playersWithStats: {
        id: number;
        name: string;
        iconClass: string;
        colorClass: string;
        wins: number;
    }[];
    ties: number;
};